Yuck...   This is the last time I do this manually.
We need to keep an explicitly $VERSION in most of these module files,
because in the older distributions of this code, it existed, and was
therefore indexed by PAUSE.  If you remove the VERSION, then CPAN will
list the older version and refer to the older distribution.

Essentially, once you have a VERSION definition, you're stuck with it.