Converted from Module::Install to EU::MM
The Makefile.PL has been converted to use EU::MM, and the
efsdeploy-dependent Version.pom automation included, for now.  When we
want to release this to CPAN (which we should do at some point) then
we'll need to figure out how to produce a packages distro that
includes the Version.pm file pre-generated, and modify Makefile.PL to
use that instead of generate it.